+++ packages/terra-framework-docs/src/terra-dev-site/test/worklist-data-grid/WorklistDataGridWithRowSelection.test.jsx
@@ -39,14 +39,15 @@ const WorklistDataGridWithRowSelection = () => {
   };
 
   const onRowSelect = (rowIdsToSelect, rowIdsToUnselect) => {
-    if (rowIdsToUnselect) {
-      rows.forEach((r, index) => { if (rowIdsToUnselect.indexOf(r.id) >= 0) { rows[index].isSelected = false; } });
-    }
+    rows.forEach((row, index) => {
+      if (rowIdsToUnselect && rowIdsToUnselect.indexOf(row.id) >= 0) {
+        rows[index].isSelected = false;
+      }
 
-    if (rowIdsToSelect) {
-      rows.forEach((r, index) => { if (rowIdsToSelect.indexOf(r.id) >= 0) { rows[index].isSelected = true; } });
-      setSelectedRows(rowIdsToSelect);
-    }
+      if (rowIdsToSelect && rowIdsToSelect.indexOf(row.id) >= 0) {
+        rows[index].isSelected = true;
+      }
+    });
   };
 
   return (