+++ packages/terra-worklist-data-grid/tests/wdio/worklist-data-grid-spec.js
@@ -1,4 +1,4 @@
-const selector = '#default-terra-worklist-data-grid';
+const defaultSelector = '#default-terra-worklist-data-grid';
 
 const holdDownShiftKey = () => {
   browser.performActions([{
@@ -32,13 +32,13 @@ const rowSelectionNavigateToCell = (row, col) => {
   moveCurrentPositionBy(row > 0 ? row - 1 : 0, col);
 };
 
-const clickCell = (row, col) => {
-  browser.$$('#default-terra-worklist-data-grid tr')[row].$(`:nth-child(${col + 1})`).click();
+const clickCell = (row, col, selector) => {
+  browser.$$(`${selector} tr`)[row].$(`:nth-child(${col + 1})`).click();
 };
 
-const shiftClickCell = (row, col) => {
+const shiftClickCell = (row, col, selector) => {
   holdDownShiftKey();
-  clickCell(row, col);
+  clickCell(row, col, selector);
   releaseKey();
 };
 
@@ -46,27 +46,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
   describe('with keyboard navigation', () => {
     beforeEach(() => {
       browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/default-worklist-data-grid');
-    });
-
-    it('tabs into the grid', () => {
-      browser.keys(['Tab']);
-
-      Terra.validates.element('first-cell-in-grid-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
-    });
-
-    it('tabs in and shift tab out of the grid', () => {
-      browser.keys(['Tab', 'Shift', 'Tab', 'Shift']);
-
-      Terra.validates.element('tab-selection-off-grid', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
-    });
-
-    it('tabs forward out of the grid', () => {
-      browser.keys(['Tab', 'Tab']);
-
-      Terra.validates.element('tab-selection-off-grid', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('uses arrow keys to navigate', () => {
@@ -77,8 +56,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
         .concat(new Array(1).fill('ArrowLeft')));
 
       browser.pause(150);
-      Terra.validates.element('arrow-key-navigation', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('arrow-key-navigation', { selector: defaultSelector });
     });
 
     it('validates that the up arrow key cannot navigate outside the grid', () => {
@@ -86,8 +64,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
         .concat(new Array(3).fill('ArrowUp')));
 
       browser.pause(150);
-      Terra.validates.element('first-cell-in-grid-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('first-cell-in-grid-focused', { selector: defaultSelector });
     });
 
     it('validates that the left arrow key cannot navigate outside the grid', () => {
@@ -95,8 +72,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
         .concat(new Array(2).fill('ArrowLeft')));
 
       browser.pause(150);
-      Terra.validates.element('first-cell-in-grid-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('first-cell-in-grid-focused', { selector: defaultSelector });
     });
 
     it('validates that Control + Home goes to first non-header cell', () => {
@@ -106,8 +82,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
         .concat(['Control', 'Home', 'Control']));
 
       browser.pause(150);
-      Terra.validates.element('first-non-column-header-cell-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('first-non-column-header-cell-focused', { selector: defaultSelector });
     });
 
     it('validates that the Home key goes to the first cell in current row', () => {
@@ -117,16 +92,14 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
         .concat(['Home']));
 
       browser.pause(150);
-      Terra.validates.element('first-cell-in-row-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('first-cell-in-row-focused', { selector: defaultSelector });
     });
 
     it('validates that the End key goes to the last cell in current row', () => {
       browser.keys(['Tab', 'ArrowDown', 'End']);
 
       browser.pause(150);
-      Terra.validates.element('last-cell-in-row-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('last-cell-in-row-focused', { selector: defaultSelector });
     });
 
     it('validates that Control + End key goes to the last cell in the grid', () => {
@@ -134,8 +107,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
         .concat(['Control', 'End', 'Control']));
 
       browser.pause(150);
-      Terra.validates.element('last-cell-in-last-row-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('last-cell-in-last-row-focused', { selector: defaultSelector });
     });
 
     it('validates that the down arrow key cannot navigate outside the grid', () => {
@@ -144,8 +116,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
         .concat(new Array(2).fill('ArrowDown')));
 
       browser.pause(150);
-      Terra.validates.element('last-cell-in-last-row-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('last-cell-in-last-row-focused', { selector: defaultSelector });
     });
 
     it('validates that the right arrow key cannot navigate outside the grid', () => {
@@ -154,8 +125,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
         .concat(new Array(2).fill('ArrowRight')));
 
       browser.pause(150);
-      Terra.validates.element('last-cell-in-last-row-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('last-cell-in-last-row-focused', { selector: defaultSelector });
     });
   });
 
@@ -168,16 +138,14 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       navigateToCell(1, 1);
       browser.keys(['Space']);
 
-      Terra.validates.element('selection-and-focus-on-1-1', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('selection-and-focus-on-1-1', { selector: defaultSelector });
     });
 
     it('validates that selecting an already selected cell has no effect', () => {
       navigateToCell(1, 1);
       browser.keys(['Space', 'Space']);
 
-      Terra.validates.element('selection-and-focus-on-1-1', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('selection-and-focus-on-1-1', { selector: defaultSelector });
     });
 
     it('validates that navigating the grid does not change the selected cell', () => {
@@ -185,8 +153,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Space']);
       moveCurrentPositionBy(2, 0);
 
-      Terra.validates.element('selection-1-1-focus-3-1', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('selection-1-1-focus-3-1', { selector: defaultSelector });
     });
 
     it('validates that selecting another cell moves the selection and focus to current cell', () => {
@@ -195,39 +162,34 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(2, 0);
       browser.keys(['Space']);
 
-      Terra.validates.element('selection-3-1-focus-3-1', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('selection-3-1-focus-3-1', { selector: defaultSelector });
     });
 
     it('validates that clicking on a cell, moves selection and focus to that cell', () => {
-      clickCell(3, 1);
+      clickCell(3, 1, defaultSelector);
 
-      Terra.validates.element('selection-3-1-focus-3-1-with-hover', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('selection-3-1-focus-3-1-with-hover', { selector: defaultSelector });
     });
 
     it('validates that navigation does not change cell selected via mouse click', () => {
-      clickCell(3, 1);
+      clickCell(3, 1, defaultSelector);
       moveCurrentPositionBy(-2, -1);
 
-      Terra.validates.element('selection-3-1-focus-1-0', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('selection-3-1-focus-1-0', { selector: defaultSelector });
     });
 
     it('validates that mouse click moves selection and focus to clicked cell', () => {
-      clickCell(3, 1); // Selection and Focus is on (3,1)
-      clickCell(4, 2); // Selection and Focus should move to (4,2)
+      clickCell(3, 1, defaultSelector); // Selection and Focus is on (3,1)
+      clickCell(4, 2, defaultSelector); // Selection and Focus should move to (4,2)
 
-      Terra.validates.element('selection-4-2-focus-4-2', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('selection-4-2-focus-4-2', { selector: defaultSelector });
     });
 
     it('clears selection with Esc key', () => {
-      clickCell(3, 1);
+      clickCell(3, 1, defaultSelector);
       browser.keys(['Escape']);
 
-      Terra.validates.element('cell-3-1-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('cell-3-1-focused', { selector: defaultSelector });
     });
   });
 
@@ -239,32 +201,28 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
     it('verifies that first tab into grid focuses on the checkbox', () => {
       browser.keys(['Tab']);
 
-      Terra.validates.element('row-1-checkbox-focused-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('row-1-checkbox-focused-row-selection-mode', { selector: defaultSelector });
     });
 
     it('navigates to last cell with Control + End key', () => {
       rowSelectionNavigateToCell(4, 3);
       browser.keys(['Control', 'End', 'Control']);
 
-      Terra.validates.element('last-cell-focused-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('last-cell-focused-row-selection-mode', { selector: defaultSelector });
     });
 
     it('navigates to first non-header cell with Control + Home key', () => {
       rowSelectionNavigateToCell(4, 3);
       browser.keys(['Control', 'Home', 'Control']);
 
-      Terra.validates.element('row-1-checkbox-focused-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('row-1-checkbox-focused-row-selection-mode', { selector: defaultSelector });
     });
 
     it('selects the entire row with spacebar', () => {
       rowSelectionNavigateToCell(4, 3);
       browser.keys(['Space']);
 
-      Terra.validates.element('row-4-selected-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('row-4-selected-row-selection-mode', { selector: defaultSelector });
     });
 
     it('unselects an already selected row with spacebar', () => {
@@ -277,8 +235,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(-3, -2);
       browser.keys(['Space']); // unselect row 1
 
-      Terra.validates.element('row-4-selected-after-unselect-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('row-4-selected-after-unselect-row-selection-mode', { selector: defaultSelector });
     });
 
     it('can still navigate when a row is selected', () => {
@@ -286,16 +243,14 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Space']); // Row 1 now selected
       moveCurrentPositionBy(3, 2); // Cell (4,5)
 
-      Terra.validates.element('row-1-selected-cell-4-5-focused-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('row-1-selected-cell-4-5-focused-row-selection-mode', { selector: defaultSelector });
     });
 
     it('selects all rows with Control + A', () => {
       rowSelectionNavigateToCell(3, 2);
       browser.keys(['Control', 'A', 'Control']);
 
-      Terra.validates.element('all-rows-selected-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('all-rows-selected-row-selection-mode', { selector: defaultSelector });
     });
 
     it('clears all selected rows with Esc', () => {
@@ -305,31 +260,27 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Space']); // Row 2 selected
       browser.keys(['Escape']); // clear both selected rows
 
-      Terra.validates.element('no-rows-selected-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('no-rows-selected-row-selection-mode', { selector: defaultSelector });
     });
 
     it('clears row selection mode when no rows are selected', () => {
       rowSelectionNavigateToCell(4, 1);
       browser.keys(['Escape']); // clears row selection mode
 
-      Terra.validates.element('row-selection-mode-off-cell-4-1-focused', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('row-selection-mode-off-cell-4-1-focused', { selector: defaultSelector });
     });
 
     it('selects a row by clicking inside that row', () => {
-      clickCell(3, 1);
+      clickCell(3, 1, defaultSelector);
 
-      Terra.validates.element('row-3-selected-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('row-3-selected-row-selection-mode', { selector: defaultSelector });
     });
 
     it('unselects a selected row by clicking in that row', () => {
-      clickCell(3, 2);
-      clickCell(3, 4);
+      clickCell(3, 2, defaultSelector);
+      clickCell(3, 4, defaultSelector);
 
-      Terra.validates.element('no-rows-selected-after-click-row-selection-mode', { selector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
+      Terra.validates.element('no-rows-selected-after-click-row-selection-mode', { selector: defaultSelector });
     });
 
     it('hovers over selectable cell in worklist data grid', () => {
@@ -353,7 +304,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       rowSelectionNavigateToCell(3, 0);
       browser.keys(['Shift', 'Space', 'Shift']);
 
-      Terra.validates.element('row-3-selected-without-hover-row-selection-mode', '#default-terra-worklist-data-grid');
+      Terra.validates.element('row-3-selected-without-hover-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
@@ -363,7 +314,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(2, -1);
       browser.keys(['Shift', 'Space', 'Shift']);
 
-      Terra.validates.element('row-3-to-5-selected-row-selection-mode', { selector });
+      Terra.validates.element('row-3-to-5-selected-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
@@ -373,7 +324,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(-4, 2);
       browser.keys(['Shift', 'Space', 'Shift']);
 
-      Terra.validates.element('row-7-to-3-selected-row-selection-mode', { selector });
+      Terra.validates.element('row-7-to-3-selected-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
@@ -385,7 +336,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(5, 1);
       browser.keys(['Shift', 'Space', 'Shift']);
 
-      Terra.validates.element('row-5-to-8-selected-row-selection-mode', { selector });
+      Terra.validates.element('row-5-to-8-selected-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
@@ -399,7 +350,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(2, 2);
       browser.keys(['Shift', 'Space', 'Shift']);
 
-      Terra.validates.element('row-multiple-ranges-selected-row-selection-mode', { selector });
+      Terra.validates.element('row-multiple-ranges-selected-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
@@ -408,7 +359,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       rowSelectionNavigateToCell(3, 2);
       browser.keys(['Shift', 'ArrowUp', 'Shift']);
 
-      Terra.validates.element('row-3-to-2-selected-row-selection-mode', { selector });
+      Terra.validates.element('row-3-to-2-selected-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
@@ -417,7 +368,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       rowSelectionNavigateToCell(3, 0);
       browser.keys(['Shift', 'ArrowDown', 'ArrowDown', 'Shift']);
 
-      Terra.validates.element('row-3-to-5-selected-row-selection-mode', { selector });
+      Terra.validates.element('row-3-to-5-selected-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
@@ -427,7 +378,7 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(4, 1);
       browser.keys(['Shift', 'ArrowDown', 'ArrowDown', 'Shift']);
 
-      Terra.validates.element('multiple-ranges-selected-row-selection-mode', { selector });
+      Terra.validates.element('multiple-ranges-selected-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
@@ -435,49 +386,49 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       rowSelectionNavigateToCell(5, 1);
       browser.keys(['Shift', 'ArrowUp', 'ArrowUp', 'ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowDown', 'Shift']);
 
-      Terra.validates.element('moving-towards-anchor-unselects-rows-row-selection-mode', { selector });
+      Terra.validates.element('moving-towards-anchor-unselects-rows-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('enables row selection mode by Shift+Click when turned off', () => {
       browser.keys(['Tab', 'Escape']);
-      clickCell(3, 3);
-      shiftClickCell(5, 1);
+      clickCell(3, 3, defaultSelector);
+      shiftClickCell(5, 1, defaultSelector);
 
-      Terra.validates.element('row-3-to-5-selected-with-hover-row-selection-mode', { selector });
+      Terra.validates.element('row-3-to-5-selected-with-hover-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('select rows upwards of anchor row by Shift+Click', () => {
-      clickCell(6, 4);
-      shiftClickCell(4, 2);
+      clickCell(6, 4, defaultSelector);
+      shiftClickCell(4, 2, defaultSelector);
 
-      Terra.validates.element('row-4-to-6-selected-with-hover-row-selection-mode', { selector });
+      Terra.validates.element('row-4-to-6-selected-with-hover-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('select multiple batches of rows by Shift+Click', () => {
-      clickCell(5, 2);
-      shiftClickCell(3, 1);
+      clickCell(5, 2, defaultSelector);
+      shiftClickCell(3, 1, defaultSelector);
 
-      clickCell(7, 2);
-      shiftClickCell(9, 3);
+      clickCell(7, 2, defaultSelector);
+      shiftClickCell(9, 3, defaultSelector);
 
-      Terra.validates.element('multiple-ranges-selected-with-hover-row-selection-mode', { selector });
+      Terra.validates.element('multiple-ranges-selected-with-hover-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('unselects rows when moving towards or across anchor row', () => {
-      clickCell(5, 3);
+      clickCell(5, 3, defaultSelector);
 
       holdDownShiftKey();
 
-      clickCell(3, 4);
-      clickCell(9, 1);
+      clickCell(3, 4, defaultSelector);
+      clickCell(9, 1, defaultSelector);
 
       releaseKey();
 
-      Terra.validates.element('row-5-to-9-selected-with-hover-row-selection-mode', { selector });
+      Terra.validates.element('row-5-to-9-selected-with-hover-row-selection-mode', { selector: defaultSelector });
       expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
   });
@@ -489,60 +440,118 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/worklist-data-grid-focusable-cell');
     });
 
+    it('validates that the proper grid element is selected when Tab is used to give initial focus to the grid', () => {
+      browser.keys(['Tab', 'Tab']);
+
+      Terra.validates.element('worklist-initial-focus', { columnResizeSelector });
+      expect(browser.$('[role="grid"] thead tr:nth-of-type(1) th:nth-of-type(1)').isFocused());
+    });
+
+    it('validates that a Tab key press inside the grid will skip focusable cell elements', () => {
+      browser.keys(['Tab', 'Tab', 'Tab']);
+
+      Terra.validates.element('worklist-skip-focusable-elements-next', { columnResizeSelector });
+      expect(browser.$('#next-focus-button').isFocused());
+    });
+
+    it('validates that a Shift+Tab key press while inside the grid will skip to the previous focusable element outside the grid', () => {
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowDown', 'Shift', 'Tab', 'Shift']);
+
+      Terra.validates.element('worklist-skip-focusable-elements-previous', { columnResizeSelector });
+      expect(browser.$('#previous-focus-button').isFocused());
+    });
+
+    it('validates that the proper element is selected when Shift+Tab is used to give focus to the grid', () => {
+      browser.keys(['Tab', 'Tab', 'Tab', 'Shift', 'Tab', 'Shift']);
+
+      Terra.validates.element('worklist-return-focus', { columnResizeSelector });
+      expect(browser.$('[role="grid"] thead tr:nth-of-type(1) th:nth-of-type(1)').isFocused());
+    });
+
     it('validates that a cell with no focusable elements does not trap focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('non-focusable-cell-no-trap', { columnResizeSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with a button element traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-button-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('button:focus')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that Escape can be used to release a focus trap', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'Enter', 'Escape']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'Enter', 'Escape']);
 
       Terra.validates.element('focusable-button-cell-escape-trap', { columnResizeSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with an input element traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-input-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('input:focus')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with an anchor element with href traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-anchor-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('a:focus')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with multiple focusable elements traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-multiple-element-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('button:focus')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with a select element traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowLeft']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowLeft']);
 
       Terra.validates.element('focusable-select-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('select:focus')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with a text area element traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-textarea-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('textarea:focus')).toBeElementsArrayOfSize(1);
+    });
+  });
+
+  describe('with pinned columns', () => {
+    const pinnedColumnsSelector = '#pinned-columns';
+    const pinnedColumnsWithRowSelectionSelector = '#pinned-columns-with-row-selection';
+
+    it('adds pinned columns', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/pinned-columns');
+
+      Terra.validates.element('pinned-columns', { selector: pinnedColumnsSelector });
+    });
+
+    it('keeps pinned columns stickied when scrolling', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/pinned-columns');
+      navigateToCell(2, 1);
+      moveCurrentPositionBy(0, 9);
+      Terra.validates.element('pinned-columns-scroll', { selector: pinnedColumnsSelector });
+    });
+
+    it('keeps row selection column stickied when scrolling', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/pinned-columns-with-row-selection');
+      navigateToCell(2, 1);
+      moveCurrentPositionBy(0, 9);
+      Terra.validates.element('pinned-columns-with-row-selection-scroll', { selector: pinnedColumnsWithRowSelectionSelector });
+    });
+
+    it('selects the whole row with pinned columns', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/pinned-columns-with-row-selection');
+
+      clickCell(2, 0, pinnedColumnsWithRowSelectionSelector);
+      Terra.validates.element('pinned-columns-with-row-selection-select', { selector: pinnedColumnsWithRowSelectionSelector });
     });
   });
 });