+++ packages/terra-tabs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 7.4.0 - (August 10, 2023)
+
+* Changed
+   * Added drag and drop to tab menu and internationalization support for the same.
+   * Updated styles for tab content panel.
+
 ## 7.3.1 - (August 4, 2023)
 
 * Changed